--- src/main/java/com/pathbook/pathbook_api/SecurityConfig.java
@@ -1,53 +1,0 @@
-package com.pathbook.pathbook_api;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy; // ✅ 추가
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-                .csrf(csrf -> csrf.disable()) // CSRF 비활성화 (필요 시 활성화)
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/auth/login", "/auth/register").permitAll() // 로그인 및 회원가입은 인증 없이 접근 가능
-                        .anyRequest().authenticated() // 그 외 모든 요청은 인증 필요
-                )
-                .formLogin(login -> login // 기본 로그인 폼 활성화 (API 방식 사용 가능)
-                        .loginPage("/auth/login") // 로그인 페이지 (프론트엔드에서 API 호출)
-                        .defaultSuccessUrl("/home", true) // 로그인 성공 후 이동할 기본 페이지
-                        .permitAll()
-                )
-                .logout(logout -> logout // 로그아웃 설정
-                        .logoutUrl("/auth/logout")
-                        .logoutSuccessUrl("/auth/login?logout")
-                        .invalidateHttpSession(true)
-                        .deleteCookies("JSESSIONID")
-                        .permitAll()
-                )
-                .sessionManagement(session -> session // 세션 기반 인증 활성화
-                        .sessionCreationPolicy(SessionCreationPolicy.IF_REQUIRED) // ✅ 정상 작동
-                );
-
-        return http.build();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder(); // 비밀번호 암호화를 위한 BCryptPasswordEncoder 사용
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration)
-            throws Exception {
-        return authenticationConfiguration.getAuthenticationManager(); // AuthenticationManager 설정
-    }
-}
