+++ src/main/java/com/pathbook/pathbook_api/controller/AuthController.java
@@ -1,14 +1,18 @@
 package com.pathbook.pathbook_api.controller;
 
-import jakarta.servlet.http.HttpSession;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
-import com.pathbook.pathbook_api.service.AuthService;
 import com.pathbook.pathbook_api.entity.User;
+import com.pathbook.pathbook_api.service.AuthService;
 
-import java.util.Map;
+import jakarta.servlet.http.HttpSession;
 
 @RestController
 @RequestMapping("/auth")
@@ -20,7 +24,6 @@ public class AuthController {
         this.authService = authService;
     }
 
-    // 로그인 (세션 방식)
     @PostMapping("/login")
     public ResponseEntity<String> login(@RequestBody User user, HttpSession session) {
         boolean success = authService.login(user.getEmail(), user.getPassword(), session);
@@ -31,14 +34,12 @@ public class AuthController {
         }
     }
 
-    // 회원가입
     @PostMapping("/register")
     public ResponseEntity<String> register(@RequestBody User user) {
         String response = authService.register(user.getUsername(), user.getEmail(), user.getPassword());
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    // 이메일 인증
     @GetMapping("/verify")
     public ResponseEntity<String> verifyEmail(@RequestParam String verificationToken) {
         boolean verified = authService.verifyEmail(verificationToken);
@@ -48,4 +49,5 @@ public class AuthController {
             return new ResponseEntity<>("이메일 인증 실패.", HttpStatus.UNAUTHORIZED);
         }
     }
+
 }
