+++ src/main/java/com/pathbook/pathbook_api/repository/UserRepository.java
@@ -15,7 +15,6 @@ public interface UserRepository extends CrudRepository<User, String> {
 
     User findByEmail(String email);
 
-    // ✅ 이메일 인증을 위한 메서드 추가
     User findByVerificationToken(String verificationToken);
 
     // TODO: DB 커넥션 체크용, 프로덕션에서는 지워야 함.
